A few optimizations - compare strings if headers have single value - support for blocking writes when queue is disabled
