+++ common/socket/src/main/java/io/helidon/common/socket/SocketWriterDirect.java
@@ -1,0 +1,45 @@
+/*
+ * Copyright (c) 2022 Oracle and/or its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.helidon.common.socket;
+
+import io.helidon.common.buffers.BufferData;
+import io.helidon.common.buffers.DataWriter;
+
+/**
+ * Socket writer (possibly) used from multiple threads, takes care of writing to a single
+ * socket.
+ */
+class SocketWriterDirect extends SocketWriter implements DataWriter {
+    /**
+     * A new socket writer.
+     *
+     * @param socket socket to write to
+     */
+    SocketWriterDirect(HelidonSocket socket) {
+        super(socket);
+    }
+
+    @Override
+    public void write(BufferData... buffers) {
+        writeNow(buffers);
+    }
+
+    @Override
+    public void write(BufferData buffer) {
+        writeNow(buffer);
+    }
+}
