+++ nima/tests/benchmark/techempower/src/main/resources/application.yaml
@@ -22,6 +22,8 @@ server:
       backlog: 8192
       receive-buffer-size: 64000
       write-queue-length: 8192
+      # use the following to disable asynchronous
+      # write-queue-length: 0
       connection-options:
         read-timeout-seconds: 0
         connect-timeout-seconds: 0
