+++ nima/webserver/webserver/src/main/java/io/helidon/nima/webserver/ServerListener.java
@@ -169,6 +169,13 @@ class ServerListener {
                                            connectedPort,
                                            socketName));
 
+            if (listenerConfig.writeQueueLength() <= 1) {
+                LOGGER.log(System.Logger.Level.INFO, "[" + serverChannelId + "] direct writes");
+            } else {
+                LOGGER.log(System.Logger.Level.INFO,
+                           "[" + serverChannelId + "] async writes, queue length: " + listenerConfig.writeQueueLength());
+            }
+
             if (LOGGER.isLoggable(TRACE)) {
                 if (listenerConfig.hasTls()) {
                     debugTls(serverChannelId, listenerConfig.tls());
